+++ internal/sit/sit.go
@@ -548,14 +548,14 @@ func readerFor(algo algid, unpacksz uint32, r io.Reader) io.Reader { // might al
 	case 0: // no compression
 		return r
 	case 3: // Huffman compression
-		return huffman(r, int64(unpacksz))
+		return huffman(r, unpacksz)
 	// case 1: // RLE compression
 	// case 2: // LZC compression
 	// case 5: // LZ with adaptive Huffman
 	// case 6: // Fixed Huffman table
 	// case 8: // Miller-Wegman encoding
 	case 13: // anonymous
-		return sit13(r, int64(unpacksz))
+		return sit13(r, unpacksz)
 	// case 14: // anonymous
 	case 15: // Arsenic
 		return arsenic(r, unpacksz)
